Refactor building query
